win32 compat

+++ audio/src/audio.cpp
@@ -62,7 +62,7 @@ unsigned int AudioImp::load_music(string fname)
   if(music)
   {
     music_tracks.push_back(music);
-    return music_tracks.size() - 1;
+    return (unsigned int)music_tracks.size() - 1;
   }
   cout<<"Mix_LoadMus: "<<Mix_GetError()<<endl;
   return -1;
@@ -82,7 +82,7 @@ unsigned int AudioImp::load_sample(string fname)
   if(sample)
   {
     samples.push_back(sample);
-    return samples.size() - 1;
+    return (unsigned int)samples.size() - 1;
   }
 
   cout<<"Mix_LoadWAV: "<<Mix_GetError()<<endl;
